requirements.txt
itchat==1.2.30
Pillow==4.0.0
jupyter-client==5.0.1
jupyter-core==4.3.0